+++ jpos/src/main/java/org/jpos/iso/channel/PostChannel.java
@@ -79,9 +79,8 @@ public class PostChannel extends BaseChannel {
     protected int getMessageLength() throws IOException, ISOException {
         byte[] b = new byte[2];
         serverIn.readFully(b,0,2);
-        return (int) (
-            ((((int)b[0])&0xFF) << 8) | 
-            (((int)b[1])&0xFF));
+        return ((((int)b[0])&0xFF) << 8) |
+        (((int)b[1])&0xFF);
     }
    /**
     *      * @param header Hex representation of header
